move shell to demo group, create file-upload service
