+++ libs/demo/shell/feature/src/lib/home.component.scss
