+++ libs/demo/shell/feature/src/lib/home.component.spec.ts
