+++ libs/file-upload/feature/src/lib/file-upload.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { FileUploadService } from './file-upload.service';
+
+describe('FileUploadService', () => {
+  let service: FileUploadService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(FileUploadService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
