+++ libs/file-upload/feature/src/lib/file-upload.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class FileUploadService {
+
+  constructor() { }
+}
